Create one function to handle follow/unfollow
